requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.7
django-image-uploader-widget==0.2.3
django_debug_toolbar==3.8.1
fake-useragent==1.1.1
idna==3.4
Pillow==9.4.0
psycopg2==2.9.5
python-dotenv==0.21.1
requests==2.28.2
soupsieve==2.4
sqlparse==0.4.3
urllib3==1.26.14
